+++ src/features/chat_etiquette/feature.py
@@ -1,6 +1,6 @@
 from features.base import Feature
 from features.chat_etiquette.details import ChatEtiquetteDetails
-from features.chat_etiquette.handlers import NameShopTalk
+from features.chat_etiquette.handlers import NameShopTalk, RemindOfEtiquette
 from features.journal import Journal
 
 
@@ -9,3 +9,4 @@ class ChatEtiquette(Feature):
 
     def register(self, journal: Journal) -> None:
         journal.events.handler(NameShopTalk())
+        journal.events.handler(RemindOfEtiquette())